services/web/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.4
flask_marshmallow==0.11.0
marshmallow-sqlalchemy==0.22.3
Flask-Migrate==2.5.2
PyGithub
ptvsd == 4.3.2